refactor(http): extract span URL redaction helper
